+++ decomposed/src/FFT_CachedTwiddles.ipp
@@ -34,7 +34,7 @@ void fft_ensure_table(int k){
         fft_ensure_table(k - 1);
     }
 
-    size_t length = 1 << k;
+    size_t length = (size_t)1 << k;
     double omega = 2 * PI / length;
     length /= 2;
 
@@ -70,7 +70,7 @@ void fft_forward(complex<double> *T, int k){
         return;
     }
 
-    size_t length = 1 << k;
+    size_t length = (size_t)1 << k;
     size_t half_length = length / 2;
 
     //  Get local twiddle table.
@@ -116,7 +116,7 @@ void fft_inverse(complex<double> *T, int k){
         return;
     }
 
-    size_t length = 1 << k;
+    size_t length = (size_t)1 << k;
     size_t half_length = length / 2;
 
     //  Recursively perform FFT on lower elements.
@@ -149,7 +149,7 @@ void fft_pointwise(complex<double> *T, const complex<double> *A, int k){
     //  -   T           -   Pointer to array.
     //  -   k           -   2^k is the size of the transform
 
-    size_t length = 1 << k;
+    size_t length = (size_t)1 << k;
     for (size_t c = 0; c < length; c++){
         T[c] = T[c] * A[c];
     }
@@ -163,7 +163,7 @@ void int_to_fft(complex<double> *T, int k, const uint32_t *A, size_t AL){
     //  -   A   -   word array
     //  -   AL  -   length of word array
 
-    size_t fft_length = 1 << k;
+    size_t fft_length = (size_t)1 << k;
     complex<double> *Tstop = T + fft_length;
 
     //  Since there are 9 digits per word and we want to put 3 digits per
@@ -196,7 +196,7 @@ void fft_to_int(const complex<double> *T, int k, uint32_t *A, size_t AL){
     //  -   AL  -   length of word array
 
     //  Compute Scaling Factor
-    size_t fft_length = 1 << k;
+    size_t fft_length = (size_t)1 << k;
     double scale = 1. / fft_length;
 
     //  Since there are 9 digits per word and we want to put 3 digits per
