+++ decomposed/src/FFT_SSE3.ipp
@@ -44,7 +44,7 @@ void fft_ensure_table(int k){
         fft_ensure_table(k - 1);
     }
 
-    size_t length = 1 << k;
+    size_t length = (size_t)1 << k;
     double omega = 2 * PI / length;
     length /= 2;
 
@@ -80,7 +80,7 @@ void fft_forward(__m128d *T, int k){
         return;
     }
 
-    size_t length = 1 << k;
+    size_t length = (size_t)1 << k;
     size_t half_length = length / 2;
 
     //  Get local twiddle table.
@@ -137,7 +137,7 @@ void fft_inverse(__m128d *T, int k){
         return;
     }
 
-    size_t length = 1 << k;
+    size_t length = (size_t)1 << k;
     size_t half_length = length / 2;
 
     //  Recursively perform FFT on lower elements.
@@ -182,7 +182,7 @@ void fft_pointwise(__m128d *T, __m128d *A, int k){
     //  -   T           -   Pointer to array.
     //  -   k           -   2^k is the size of the transform
 
-    size_t length = 1 << k;
+    size_t length = (size_t)1 << k;
     for (size_t c = 0; c < length; c++){
         __m128d a0 = T[c];
         __m128d b0 = A[c];
@@ -201,7 +201,7 @@ void int_to_fft(__m128d *T, int k, const uint32_t *A, size_t AL){
     //  -   A   -   word array
     //  -   AL  -   length of word array
 
-    size_t fft_length = 1 << k;
+    size_t fft_length = (size_t)1 << k;
     __m128d *Tstop = T + fft_length;
 
     //  Since there are 9 digits per word and we want to put 3 digits per
@@ -234,7 +234,7 @@ void fft_to_int(__m128d *T, int k, uint32_t *A, size_t AL){
     //  -   AL  -   length of word array
 
     //  Compute Scaling Factor
-    size_t fft_length = 1 << k;
+    size_t fft_length = (size_t)1 << k;
     double scale = 1. / fft_length;
 
     //  Since there are 9 digits per word and we want to put 3 digits per
